+++ packages/snap-account-service/CHANGELOG.md
@@ -46,6 +46,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - If the migration is not successful, it will get retried everytime we need to interact with any Snap keyring (v2) instances.
   - It is conccurent-free and can safely be called by multiple execution flows.
   - Once the migration has ran, the legacy Snap keyring will be emptied, thus, consumers are expected to use the new per-Snap keyring (v2) instances instead.
+  - Selected-account forwarding now targets v2 Snap keyrings.
+  - The service messenger now requires the `KeyringController:withKeyringV2Unsafe`.
 
 ### Changed